Moved class ElementNameUtil
- Moved class
- Updated references
